requirements.txt
accelerate==1.11.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
anyio==4.12.0
async-timeout==5.0.1
attrs==25.4.0
bitsandbytes==0.42.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.2
cycler==0.12.1
datasets==4.4.1
dill==0.4.0
exceptiongroup==1.3.1
filelock==3.20.0
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.10.0
ftfy==6.3.1
gensim==4.4.0
googletrans==4.0.2
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1765794027227/work
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
MedCLIP==0.0.3
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
networkx==3.4.2
nltk==3.9.2
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1747544633931/work/dist/numpy-2.2.6-cp310-cp310-macosx_11_0_arm64.whl#sha256=50a17562ef677de6531058e998e32487919c216e19ca4e1bdd86b0e1262251c1
open_clip_torch==3.2.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
propcache==0.4.1
psutil==7.1.3
pyarrow==22.0.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.6.2
scikit-learn @ file:///Users/runner/miniforge3/conda-bld/scikit-learn_1757406234687/work/dist/scikit_learn-1.7.2-cp310-cp310-macosx_11_0_arm64.whl#sha256=bb66f143da543e6bda8f24b531737ff6db838b0300a8c5847c06bc7b08fff666
scipy @ file:///Users/runner/miniforge3/conda-bld/scipy-split_1739790705989/work/dist/scipy-1.15.2-cp310-cp310-macosx_11_0_arm64.whl#sha256=7edcf528c303a2d0aa5afa5be176589f18a84d21a8d85f1972553ab95cab0933
six==1.17.0
smart_open==7.5.0
sympy==1.14.0
textaugment==2.0.0
textblob==0.19.0
threadpoolctl @ file:///home/conda/feedstock_root/build_artifacts/threadpoolctl_1741878222898/work
timm==1.0.22
tokenizers==0.22.1
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.5.0
wcwidth==0.2.14
wget==3.2
wrapt==2.0.1
xxhash==3.6.0
yarl==1.22.0
